requirements.txt
anyio==3.7.0
certifi==2023.5.7
exceptiongroup==1.1.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.2
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
mypy==0.981
mypy-extensions==0.4.3
Pillow==9.5.0
pip==23.1.2
psutil==5.9.0
python-magic-bin==0.4.14
python-telegram-bot==20.3
setuptools==67.8.0
sniffio==1.3.0
toml==0.10.2
typing_extensions==4.6.3
wheel==0.38.4
